claude-rag/pyproject.toml
[project]
name = "claude-rag"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "sentence-transformers>=3.3.0",
    "watchdog>=4.0.0",
    "tiktoken>=0.7.0",
    "mcp>=1.0.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4.0", "mypy>=1.10"]
api = ["httpx>=0.27.0"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests that load ML models (deselect with '-m \"not slow\"')",
]
